Agregados mas sensores

+++ src/master_plan.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 struct Stage {
-  Stage(int, std::string, int, double, int);
+  Stage(int, std::string, double, double, double);
   virtual ~Stage();
   std::string Serialize();
 
@@ -12,9 +12,9 @@ struct Stage {
 
   int _natural_order;
   std::string _friendly_name;
-  int _humidity_required;
+  double _humidity_required;
   double _ph_required;
-  int _temperature_required;
+  double _temperature_required;
 };
 
 class MasterPlan{
